requirements.txt
blinker==1.9.0
click==8.1.8
flask==3.1.0
importlib-metadata==8.6.1
itsdangerous==2.2.0
jinja2==3.1.6
MarkupSafe==3.0.2
python-dotenv==1.1.0
werkzeug==3.1.3
zipp==3.21.0
